+++ src/main/java/com/academy/mudogroupware/timetable/application/usecase/ExportTimetableUseCase.java
@@ -1,0 +1,8 @@
+package com.academy.mudogroupware.timetable.application.usecase;
+
+import com.academy.mudogroupware.timetable.application.command.ExportTimetableCommand;
+
+public interface ExportTimetableUseCase {
+
+    byte[] export(ExportTimetableCommand command);
+}